+++ src/ff7_opengl.cpp
@@ -420,9 +420,9 @@ void ff7_init_hooks(struct game_obj *_game_object)
 	replace_call(ff7_externals.credits_main_loop + 0xAC, ff7_credits_loop_gfx_begin_scene);
 
 	//######################
-	// snowboard .P model vertices limit fix
+	// snowboard .P model vertices limit fix + allow float vertex data type
 	//######################
-	memset_code(ff7_externals.snowboard_parse_model_vertices_732159 + 0x7E, 0x90, 6);
+	replace_function(ff7_externals.snowboard_parse_model_vertices_732159, ff7_snowboard_parse_model_vertices);
 }
 
 struct ff7_gfx_driver *ff7_load_driver(void* _game_object)