+++ src/patch.h
@@ -71,3 +71,5 @@ void patch_divide_code(uint32_t offset, int multiplier)
 
 void memcpy_code(uint32_t offset, void *data, uint32_t size);
 void memset_code(uint32_t offset, uint32_t val, uint32_t size);
+
+void* member_func_to_ptr(char i, ...);